Third addition
Rewrite Board to use two loops to make the squares instead of hardcoding them.

Holy Moley this took a lot longer than the previous two additions. I already knew that the answer required a double for loop, but I struggled a ton trying to create <div> objects in a for loop.

I eventually found a stackoverflow that showed that the render function doesn't need to take a single object; a list is fine. More importantly, it had a snippet of code that utilized pushes to an array, which worked so much easier.

This revision also allows the grid to be resized, which paves the way for other games to be created.